test(runtime): extract fork safety script
